refactor: split proposals-as-blocks.js into multiple classes
